@@ -1,6 +1,9 @@
 from distutils.core import setup
+from distutils import ccompiler
+from distutils.errors import CCompilerError, DistutilsPlatformError
 from pathlib import Path
 import re
+import tempfile
 
 import os
 import sys
@@ -21,6 +24,20 @@ def get_package_list():
     return core
 
 
+def check_compiler_available():
+    m = ccompiler.new_compiler()
+
+    with tempfile.TemporaryDirectory() as tdir:
+        with tempfile.NamedTemporaryFile(prefix="dummy", suffix=".c", dir=tdir) as tfile:
+            tfile.write(b"int main(int argc, char** argv) {return 0;}")
+            tfile.seek(0)
+            try:
+                m.compile([tfile.name])
+            except (CCompilerError, DistutilsPlatformError):
+                return False
+    return True
+
+
 def get_requirements():
     with open("requirements.txt") as f:
         requirements = f.read().splitlines()
@@ -31,6 +48,9 @@ def get_requirements():
     except ValueError:
         pass
 
+    if not check_compiler_available():  # Can't compile python-Levensthein, so drop extra
+        requirements.remove("fuzzywuzzy[speedup]<=0.16.0")
+        requirements.append("fuzzywuzzy<=0.16.0")
     if IS_DEPLOYING or not (IS_TRAVIS or IS_RTD):
         requirements.append("discord.py>=1.0.0a0")
     if sys.platform.startswith("linux"):
@@ -104,7 +124,7 @@ setup(
             "redbot-launcher=redbot.launcher:main",
         ]
     },
-    python_requires=">=3.5,<3.7",
+    python_requires=">=3.6,<3.7",
     setup_requires=get_requirements(),
     install_requires=get_requirements(),
     dependency_links=dep_links,
@@ -113,6 +133,6 @@ setup(
         "mongo": ["motor"],
         "docs": ["sphinx>=1.7", "sphinxcontrib-asyncio", "sphinx_rtd_theme"],
         "voice": ["red-lavalink>=0.0.4"],
-        "style": ["black"],
+        "style": ["black==18.5b1"],
     },
 )
