@@ -13,7 +13,7 @@ from redbot.core.events import init_events
 from redbot.core.cli import interactive_config, confirm, parse_cli_flags, ask_sentry
 from redbot.core.core_commands import Core
 from redbot.core.dev_commands import Dev
-from redbot.core import rpc, __version__
+from redbot.core import __version__
 import asyncio
 import logging.handlers
 import logging
@@ -40,7 +40,7 @@ def init_loggers(cli_flags):
     logger = logging.getLogger("red")
 
     red_format = logging.Formatter(
-        "%(asctime)s %(levelname)s %(module)s %(funcName)s %(lineno)d: " "%(message)s",
+        "%(asctime)s %(levelname)s %(module)s %(funcName)s %(lineno)d: %(message)s",
         datefmt="[%d/%m/%Y %H:%M]",
     )
 
@@ -111,7 +111,7 @@ def main():
         sys.exit(1)
     load_basic_configuration(cli_flags.instance_name)
     log, sentry_log = init_loggers(cli_flags)
-    red = Red(cli_flags, description=description, pm_help=None)
+    red = Red(cli_flags=cli_flags, description=description, pm_help=None)
     init_global_checks(red)
     init_events(red, cli_flags)
     red.add_cog(Core(red))
@@ -123,7 +123,7 @@ def main():
     loop.run_until_complete(_get_prefix_and_token(red, tmp_data))
     token = os.environ.get("RED_TOKEN", tmp_data["token"])
     prefix = cli_flags.prefix or tmp_data["prefix"]
-    if token is None or not prefix:
+    if not (token and prefix):
         if cli_flags.no_prompt is False:
             new_token = interactive_config(red, token_set=bool(token), prefix_set=bool(prefix))
             if new_token:
@@ -132,20 +132,22 @@ def main():
             log.critical("Token and prefix must be set in order to login.")
             sys.exit(1)
     loop.run_until_complete(_get_prefix_and_token(red, tmp_data))
+
+    if cli_flags.dry_run:
+        loop.run_until_complete(red.http.close())
+        sys.exit(0)
     if tmp_data["enable_sentry"]:
         red.enable_sentry()
-    cleanup_tasks = True
     try:
         loop.run_until_complete(red.start(token, bot=not cli_flags.not_bot))
     except discord.LoginFailure:
-        cleanup_tasks = False  # No login happened, no need for this
         log.critical(
             "This token doesn't seem to be valid. If it belongs to "
             "a user account, remember that the --not-bot flag "
             "must be used. For self-bot functionalities instead, "
             "--self-bot"
         )
-        db_token = red.db.token()
+        db_token = loop.run_until_complete(red.db.token())
         if db_token and not cli_flags.no_prompt:
             print("\nDo you want to reset the token? (y/n)")
             if confirm("> "):
@@ -160,11 +162,13 @@ def main():
         sentry_log.critical("Fatal Exception", exc_info=e)
         loop.run_until_complete(red.logout())
     finally:
-        red.rpc.close()
-        if cleanup_tasks:
-            pending = asyncio.Task.all_tasks(loop=red.loop)
-            gathered = asyncio.gather(*pending, loop=red.loop, return_exceptions=True)
-            gathered.cancel()
+        pending = asyncio.Task.all_tasks(loop=red.loop)
+        gathered = asyncio.gather(*pending, loop=red.loop, return_exceptions=True)
+        gathered.cancel()
+        try:
+            red.rpc.server.close()
+        except AttributeError:
+            pass
 
         sys.exit(red._shutdown_mode.value)
 
