@@ -20,7 +20,6 @@ __author__ = ["aikaterna", "billy/bollo/ati"]
 
 @cog_i18n(_)
 class Audio:
-
     def __init__(self, bot):
         self.bot = bot
         self.config = Config.get_conf(self, 2711759130, force_registration=True)
@@ -164,12 +163,11 @@ class Audio:
                 await message_channel.send(embed=embed)
                 await player.skip()
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @commands.guild_only()
     async def audioset(self, ctx):
         """Music configuration options."""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @audioset.command()
     @checks.admin_or_permissions(manage_roles=True)
@@ -426,7 +424,11 @@ class Audio:
                 await message.add_reaction(expected[i])
 
         def check(r, u):
-            return r.message.id == message.id and u == ctx.message.author
+            return (
+                r.message.id == message.id
+                and u == ctx.message.author
+                and any(e in str(r.emoji) for e in expected)
+            )
 
         try:
             (r, u) = await self.bot.wait_for("reaction_add", check=check, timeout=10.0)
@@ -628,12 +630,11 @@ class Audio:
                 await player.play()
         await ctx.send(embed=embed)
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @commands.guild_only()
     async def playlist(self, ctx):
         """Playlist configuration options."""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @playlist.command(name="append")
     async def _playlist_append(self, ctx, playlist_name, *url):
@@ -1485,11 +1486,13 @@ class Audio:
         else:
             return False
 
-    @staticmethod
-    async def _skip_action(ctx):
+    async def _skip_action(self, ctx):
         player = lavalink.get_player(ctx.guild.id)
         if not player.queue:
-            pos, dur = player.position, player.current.length
+            try:
+                pos, dur = player.position, player.current.length
+            except AttributeError:
+                return await self._embed_msg(ctx, "There's nothing in the queue.")
             time_remain = lavalink.utils.format_time(dur - pos)
             if player.current.is_stream:
                 embed = discord.Embed(
@@ -1588,13 +1591,12 @@ class Audio:
             embed.set_footer(text="Nothing playing.")
         await ctx.send(embed=embed)
 
-    @commands.group(aliases=["llset"])
+    @commands.group(aliases=["llset"], autohelp=True)
     @commands.guild_only()
     @checks.is_owner()
     async def llsetup(self, ctx):
         """Lavalink server configuration options."""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @llsetup.command()
     async def external(self, ctx):
