@@ -58,7 +58,7 @@ class Bank:
 
     # SECTION commands
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @checks.guildowner_or_permissions(administrator=True)
     async def bankset(self, ctx: commands.Context):
         """Base command for bank settings"""
@@ -69,17 +69,15 @@ class Bank:
                 default_balance = await bank._conf.default_balance()
             else:
                 if not ctx.guild:
-                    await ctx.send_help()
                     return
                 bank_name = await bank._conf.guild(ctx.guild).bank_name()
                 currency_name = await bank._conf.guild(ctx.guild).currency()
                 default_balance = await bank._conf.guild(ctx.guild).default_balance()
 
             settings = _(
-                "Bank settings:\n\n" "Bank name: {}\n" "Currency: {}\n" "Default balance: {}" ""
+                "Bank settings:\n\nBank name: {}\nCurrency: {}\nDefault balance: {}"
             ).format(bank_name, currency_name, default_balance)
             await ctx.send(box(settings))
-            await ctx.send_help()
 
     @bankset.command(name="toggleglobal")
     @checks.is_owner()
