@@ -1,11 +1,9 @@
 import discord
-from discord.ext import commands
-from .repo_manager import RepoManager
+from redbot.core import commands
 from .installable import Installable
 
 
 class InstalledCog(commands.Converter):
-
     async def convert(self, ctx: commands.Context, arg: str) -> Installable:
         downloader = ctx.bot.get_cog("Downloader")
         if downloader is None:
