@@ -15,7 +15,7 @@ from redbot.core.utils.chat_formatting import box, pagify
 from redbot.core import commands
 
 from redbot.core.bot import Red
-from .checks import install_agreement
+from .checks import do_install_agreement
 from .converters import InstalledCog
 from .errors import CloningError, ExistingGitRepo
 from .installable import Installable
@@ -27,7 +27,6 @@ _ = Translator("Downloader", __file__)
 
 @cog_i18n(_)
 class Downloader:
-
     def __init__(self, bot: Red):
         self.bot = bot
 
@@ -50,11 +49,11 @@ class Downloader:
         if str(self.LIB_PATH) not in syspath:
             syspath.insert(1, str(self.LIB_PATH))
 
-        self._repo_manager = RepoManager(self.conf)
+        self._repo_manager = RepoManager()
 
     async def cog_install_path(self):
         """Get the current cog install path.
-        
+
         Returns
         -------
         pathlib.Path
@@ -65,7 +64,7 @@ class Downloader:
 
     async def installed_cogs(self) -> Tuple[Installable]:
         """Get info on installed cogs.
-        
+
         Returns
         -------
         `tuple` of `Installable`
@@ -78,7 +77,7 @@ class Downloader:
 
     async def _add_to_installed(self, cog: Installable):
         """Mark a cog as installed.
-        
+
         Parameters
         ----------
         cog : Installable
@@ -94,7 +93,7 @@ class Downloader:
 
     async def _remove_from_installed(self, cog: Installable):
         """Remove a cog from the saved list of installed cogs.
-        
+
         Parameters
         ----------
         cog : Installable
@@ -205,17 +204,15 @@ class Downloader:
                 )
             )
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @checks.is_owner()
     async def repo(self, ctx):
         """
         Command group for managing Downloader repos.
         """
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @repo.command(name="add")
-    @install_agreement()
     async def _repo_add(self, ctx, name: str, repo_url: str, branch: str = None):
         """
         Add a new repo to Downloader.
@@ -223,6 +220,9 @@ class Downloader:
         Name can only contain characters A-z, numbers and underscore
         Branch will default to master if not specified
         """
+        agreed = await do_install_agreement(ctx)
+        if not agreed:
+            return
         try:
             # noinspection PyTypeChecker
             repo = await self._repo_manager.add_repo(name=name, url=repo_url, branch=branch)
@@ -252,19 +252,33 @@ class Downloader:
         """
         repos = self._repo_manager.get_all_repo_names()
         repos = sorted(repos, key=str.lower)
-        joined = _("Installed Repos:\n") + "\n".join(["+ " + r for r in repos])
+        joined = _("Installed Repos:\n\n")
+        for repo_name in repos:
+            repo = self._repo_manager.get_repo(repo_name)
+            joined += "+ {}: {}\n".format(repo.name, repo.short or "")
 
         for page in pagify(joined, ["\n"], shorten_by=16):
             await ctx.send(box(page.lstrip(" "), lang="diff"))
 
-    @commands.group()
+    @repo.command(name="info")
+    async def _repo_info(self, ctx, repo_name: Repo):
+        """
+        Lists information about a single repo
+        """
+        if repo_name is None:
+            await ctx.send(_("There is no repo `{}`").format(repo_name.name))
+            return
+
+        msg = _("Information on {}:\n{}").format(repo_name.name, repo_name.description or "")
+        await ctx.send(box(msg))
+
+    @commands.group(autohelp=True)
     @checks.is_owner()
     async def cog(self, ctx):
         """
         Command group for managing installable Cogs.
         """
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @cog.command(name="install")
     async def _cog_install(self, ctx, repo_name: Repo, cog_name: str):
@@ -274,7 +288,7 @@ class Downloader:
         cog = discord.utils.get(repo_name.available_cogs, name=cog_name)  # type: Installable
         if cog is None:
             await ctx.send(
-                _("Error, there is no cog by the name of" " `{}` in the `{}` repo.").format(
+                _("Error, there is no cog by the name of `{}` in the `{}` repo.").format(
                     cog_name, repo_name.name
                 )
             )
@@ -291,7 +305,7 @@ class Downloader:
 
         if not await repo_name.install_requirements(cog, self.LIB_PATH):
             await ctx.send(
-                _("Failed to install the required libraries for" " `{}`: `{}`").format(
+                _("Failed to install the required libraries for `{}`: `{}`").format(
                     cog.name, cog.requirements
                 )
             )
@@ -366,12 +380,27 @@ class Downloader:
         """
         Lists all available cogs from a single repo.
         """
+        installed = await self.installed_cogs()
+        installed_str = ""
+        if installed:
+            installed_str = _("Installed Cogs:\n") + "\n".join(
+                [
+                    "- {}{}".format(i.name, ": {}".format(i.short) if i.short else "")
+                    for i in installed
+                    if i.repo_name == repo_name.name
+                ]
+            )
         cogs = repo_name.available_cogs
         cogs = _("Available Cogs:\n") + "\n".join(
-            ["+ {}: {}".format(c.name, c.short or "") for c in cogs]
+            [
+                "+ {}: {}".format(c.name, c.short or "")
+                for c in cogs
+                if not (c.hidden or c in installed)
+            ]
         )
-
-        await ctx.send(box(cogs, lang="diff"))
+        cogs = cogs + "\n\n" + installed_str
+        for page in pagify(cogs, ["\n"], shorten_by=16):
+            await ctx.send(box(page.lstrip(" "), lang="diff"))
 
     @cog.command(name="info")
     async def _cog_info(self, ctx, repo_name: Repo, cog_name: str):
@@ -385,10 +414,14 @@ class Downloader:
             )
             return
 
-        msg = _("Information on {}:\n{}").format(cog.name, cog.description or "")
+        msg = _("Information on {}:\n{}\n\nRequirements: {}").format(
+            cog.name, cog.description or "", ", ".join(cog.requirements) or "None"
+        )
         await ctx.send(box(msg))
 
-    async def is_installed(self, cog_name: str) -> (bool, Union[Installable, None]):
+    async def is_installed(
+        self, cog_name: str
+    ) -> Union[Tuple[bool, Installable], Tuple[bool, None]]:
         """Check to see if a cog has been installed through Downloader.
 
         Parameters
@@ -419,7 +452,7 @@ class Downloader:
             Name of the command which belongs to the cog.
         cog_installable : `Installable` or `object`
             Can be an `Installable` instance or a Cog instance.
-            
+
         Returns
         -------
         str
@@ -444,12 +477,12 @@ class Downloader:
         """Determines the cog name that Downloader knows from the cog instance.
 
         Probably.
-        
+
         Parameters
         ----------
         instance : object
             The cog instance.
-            
+
         Returns
         -------
         str
