@@ -3,9 +3,8 @@ import distutils.dir_util
 import shutil
 from enum import Enum
 from pathlib import Path
-from typing import MutableMapping, Any
+from typing import MutableMapping, Any, TYPE_CHECKING
 
-from redbot.core.utils import TYPE_CHECKING
 from .log import log
 from .json_mixins import RepoJSONMixin
 
@@ -25,7 +24,7 @@ class Installable(RepoJSONMixin):
      - Modules
      - Repo Libraries
      - Other stuff?
-     
+
     The attributes of this class will mostly come from the installation's
     info.json.
 
@@ -76,6 +75,7 @@ class Installable(RepoJSONMixin):
         self.bot_version = (3, 0, 0)
         self.min_python_version = (3, 5, 1)
         self.hidden = False
+        self.disabled = False
         self.required_cogs = {}  # Cog name -> repo URL
         self.requirements = ()
         self.tags = ()
@@ -117,7 +117,7 @@ class Installable(RepoJSONMixin):
         try:
             copy_func(src=str(self._location), dst=str(target_dir / self._location.stem))
         except:
-            log.exception("Error occurred when copying path:" " {}".format(self._location))
+            log.exception("Error occurred when copying path: {}".format(self._location))
             return False
         return True
 
@@ -146,9 +146,7 @@ class Installable(RepoJSONMixin):
                 info = json.load(f)
             except json.JSONDecodeError:
                 info = {}
-                log.exception(
-                    "Invalid JSON information file at path:" " {}".format(info_file_path)
-                )
+                log.exception("Invalid JSON information file at path: {}".format(info_file_path))
             else:
                 self._info = info
 
@@ -175,6 +173,12 @@ class Installable(RepoJSONMixin):
         except ValueError:
             hidden = False
         self.hidden = hidden
+
+        try:
+            disabled = bool(info.get("disabled", False))
+        except ValueError:
+            disabled = False
+        self.disabled = disabled
 
         self.required_cogs = info.get("required_cogs", {})
 
