@@ -15,12 +15,11 @@ class ModLog:
     def __init__(self, bot: Red):
         self.bot = bot
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @checks.guildowner_or_permissions(administrator=True)
     async def modlogset(self, ctx: commands.Context):
         """Settings for the mod log"""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @modlogset.command()
     @commands.guild_only()
@@ -35,9 +34,7 @@ class ModLog:
                 await ctx.send(_("Mod events will be sent to {}").format(channel.mention))
             else:
                 await ctx.send(
-                    _("I do not have permissions to " "send messages in {}!").format(
-                        channel.mention
-                    )
+                    _("I do not have permissions to send messages in {}!").format(channel.mention)
                 )
         else:
             try:
@@ -146,5 +143,5 @@ class ModLog:
             if case_before.moderator != author:
                 to_modify["amended_by"] = author
             to_modify["modified_at"] = ctx.message.created_at.timestamp()
-            await case_before.edit(self.bot, to_modify)
+            await case_before.edit(to_modify)
             await ctx.send(_("Reason has been updated."))
