@@ -30,7 +30,6 @@ def rnd(url):
 
 
 class TwitchCommunity:
-
     def __init__(self, **kwargs):
         self.name = kwargs.pop("name")
         self.id = kwargs.pop("id", None)
@@ -119,7 +118,6 @@ class TwitchCommunity:
 
 
 class Stream:
-
     def __init__(self, **kwargs):
         self.name = kwargs.pop("name", None)
         self.channels = kwargs.pop("channels", [])
@@ -148,7 +146,6 @@ class Stream:
 
 
 class YoutubeStream(Stream):
-
     def __init__(self, **kwargs):
         self.id = kwargs.pop("id", None)
         self._token = kwargs.pop("token", None)
@@ -183,7 +180,7 @@ class YoutubeStream(Stream):
         video_url = "https://youtube.com/watch?v={}".format(vid_data["id"])
         title = vid_data["snippet"]["title"]
         thumbnail = vid_data["snippet"]["thumbnails"]["default"]["url"]
-        channel_title = data["snippet"]["channelTitle"]
+        channel_title = vid_data["snippet"]["channelTitle"]
         embed = discord.Embed(title=title, url=video_url)
         embed.set_author(name=channel_title)
         embed.set_image(url=rnd(thumbnail))
@@ -213,7 +210,6 @@ class YoutubeStream(Stream):
 
 
 class TwitchStream(Stream):
-
     def __init__(self, **kwargs):
         self.id = kwargs.pop("id", None)
         self._token = kwargs.pop("token", None)
@@ -266,7 +262,7 @@ class TwitchStream(Stream):
         url = channel["url"]
         logo = channel["logo"]
         if logo is None:
-            logo = "https://static-cdn.jtvnw.net/" "jtv_user_pictures/xarth/404_user_70x70.png"
+            logo = "https://static-cdn.jtvnw.net/jtv_user_pictures/xarth/404_user_70x70.png"
         status = channel["status"]
         if not status:
             status = "Untitled broadcast"
@@ -288,7 +284,6 @@ class TwitchStream(Stream):
 
 
 class HitboxStream(Stream):
-
     async def is_online(self):
         url = "https://api.hitbox.tv/media/live/" + self.name
 
@@ -326,7 +321,6 @@ class HitboxStream(Stream):
 
 
 class MixerStream(Stream):
-
     async def is_online(self):
         url = "https://mixer.com/api/v1/channels/" + self.name
 
@@ -348,7 +342,7 @@ class MixerStream(Stream):
             raise APIError()
 
     def make_embed(self, data):
-        default_avatar = "https://mixer.com/_latest/assets/images/main/" "avatars/default.jpg"
+        default_avatar = "https://mixer.com/_latest/assets/images/main/avatars/default.jpg"
         user = data["user"]
         url = "https://mixer.com/" + data["token"]
         embed = discord.Embed(title=data["name"], url=url)
@@ -368,7 +362,6 @@ class MixerStream(Stream):
 
 
 class PicartoStream(Stream):
-
     async def is_online(self):
         url = "https://api.picarto.tv/v1/channel/name/" + self.name
 
@@ -390,7 +383,7 @@ class PicartoStream(Stream):
 
     def make_embed(self, data):
         avatar = rnd(
-            "https://picarto.tv/user_data/usrimg/{}/dsdefault.jpg" "".format(data["name"].lower())
+            "https://picarto.tv/user_data/usrimg/{}/dsdefault.jpg".format(data["name"].lower())
         )
         url = "https://picarto.tv/" + data["name"]
         thumbnail = data["thumbnails"]["web"]
@@ -412,5 +405,5 @@ class PicartoStream(Stream):
             data["adult"] = ""
 
         embed.color = 0x4C90F3
-        embed.set_footer(text="{adult}Category: {category} | Tags: {tags}" "".format(**data))
+        embed.set_footer(text="{adult}Category: {category} | Tags: {tags}".format(**data))
         return embed
