@@ -41,31 +41,29 @@ class Warnings:
         except RuntimeError:
             pass
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @commands.guild_only()
     @checks.guildowner_or_permissions(administrator=True)
     async def warningset(self, ctx: commands.Context):
         """Warning settings"""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @warningset.command()
     @commands.guild_only()
     async def allowcustomreasons(self, ctx: commands.Context, allowed: bool):
-        """Allow or disallow custom reasons for a warning"""
+        """Enable or Disable custom reasons for a warning"""
         guild = ctx.guild
         await self.config.guild(guild).allow_custom_reasons.set(allowed)
         await ctx.send(
-            _("Custom reasons have been {}").format(_("enabled") if allowed else _("disabled"))
+            _("Custom reasons have been {}.").format(_("enabled") if allowed else _("disabled"))
         )
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @commands.guild_only()
     @checks.guildowner_or_permissions(administrator=True)
     async def warnaction(self, ctx: commands.Context):
         """Action management"""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @warnaction.command(name="add")
     @commands.guild_only()
@@ -84,7 +82,7 @@ class Warnings:
         try:
             msg = await ctx.bot.wait_for("message", check=same_author_check, timeout=30)
         except asyncio.TimeoutError:
-            await ctx.send(_("Ok then"))
+            await ctx.send(_("Ok then."))
             return
 
         if msg.content.lower() == "y":
@@ -136,13 +134,12 @@ class Warnings:
             else:
                 await ctx.send(_("No action named {} exists!").format(action_name))
 
-    @commands.group()
+    @commands.group(autohelp=True)
     @commands.guild_only()
     @checks.guildowner_or_permissions(administrator=True)
     async def warnreason(self, ctx: commands.Context):
         """Add reasons for warnings"""
-        if ctx.invoked_subcommand is None:
-            await ctx.send_help()
+        pass
 
     @warnreason.command(name="add")
     @commands.guild_only()
@@ -161,7 +158,7 @@ class Warnings:
         async with guild_settings.reasons() as registered_reasons:
             registered_reasons.update(completed)
 
-        await ctx.send(_("That reason has been registered"))
+        await ctx.send(_("That reason has been registered."))
 
     @warnreason.command(name="del")
     @commands.guild_only()
@@ -173,7 +170,7 @@ class Warnings:
             if registered_reasons.pop(reason_name.lower(), None):
                 await ctx.tick()
             else:
-                await ctx.send(_("That is not a registered reason name"))
+                await ctx.send(_("That is not a registered reason name."))
 
     @commands.command()
     @commands.guild_only()
@@ -230,7 +227,7 @@ class Warnings:
                 await ctx.send(
                     _(
                         "Custom reasons are not allowed! Please see {} for "
-                        "a complete list of valid reasons"
+                        "a complete list of valid reasons."
                     ).format("`{}reasonlist`".format(ctx.prefix))
                 )
                 return
@@ -275,7 +272,7 @@ class Warnings:
         else:
             if not await is_admin_or_superior(self.bot, ctx.author):
                 await ctx.send(
-                    warning(_("You are not allowed to check " "warnings for other users!"))
+                    warning(_("You are not allowed to check warnings for other users!"))
                 )
                 return
             else:
@@ -336,7 +333,7 @@ class Warnings:
         try:
             msg = await ctx.bot.wait_for("message", check=same_author_check, timeout=30)
         except asyncio.TimeoutError:
-            await ctx.send(_("Ok then"))
+            await ctx.send(_("Ok then."))
             return
         try:
             int(msg.content)
@@ -349,11 +346,11 @@ class Warnings:
                 return
             to_add["points"] = int(msg.content)
 
-        await ctx.send(_("Enter a description for this reason"))
+        await ctx.send(_("Enter a description for this reason."))
         try:
             msg = await ctx.bot.wait_for("message", check=same_author_check, timeout=30)
         except asyncio.TimeoutError:
-            await ctx.send(_("Ok then"))
+            await ctx.send(_("Ok then."))
             return
         to_add["description"] = msg.content
         return to_add
