@@ -4,11 +4,19 @@ This module contains extended classes and functions which are intended to
 replace those from the `discord.ext.commands` module.
 """
 import inspect
+from typing import TYPE_CHECKING
 
 from discord.ext import commands
 
+from .errors import ConversionFailure
+from ..i18n import Translator
+
+if TYPE_CHECKING:
+    from .context import Context
 
 __all__ = ["Command", "Group", "command", "group"]
+
+_ = Translator("commands.commands", __file__)
 
 
 class Command(commands.Command):
@@ -48,6 +56,78 @@ class Command(commands.Command):
         # We don't want our help property to be overwritten, namely by super()
         pass
 
+    @property
+    def parents(self):
+        """
+        Returns all parent commands of this command.
+
+        This is a list, sorted by the length of :attr:`.qualified_name` from highest to lowest.
+        If the command has no parents, this will be an empty list.
+        """
+        cmd = self.parent
+        entries = []
+        while cmd is not None:
+            entries.append(cmd)
+            cmd = cmd.parent
+        return sorted(entries, key=lambda x: len(x.qualified_name), reverse=True)
+
+    async def do_conversion(self, ctx: "Context", converter, argument: str):
+        """Convert an argument according to its type annotation.
+
+        Raises
+        ------
+        ConversionFailure
+            If doing the conversion failed.
+
+        Returns
+        -------
+        Any
+            The converted argument.
+
+        """
+        # Let's not worry about all of this junk if it's just a str converter
+        if converter is str:
+            return argument
+
+        try:
+            return await super().do_conversion(ctx, converter, argument)
+        except commands.BadArgument as exc:
+            raise ConversionFailure(converter, argument, *exc.args) from exc
+        except ValueError as exc:
+            # Some common converters need special treatment...
+            if converter in (int, float):
+                message = _('"{argument}" is not a number.').format(argument=argument)
+                raise ConversionFailure(converter, argument, message) from exc
+
+            # We should expose anything which might be a bug in the converter
+            raise exc
+
+    def command(self, cls=None, *args, **kwargs):
+        """A shortcut decorator that invokes :func:`.command` and adds it to
+        the internal command list via :meth:`~.GroupMixin.add_command`.
+        """
+        cls = cls or self.__class__
+
+        def decorator(func):
+            result = command(*args, **kwargs)(func)
+            self.add_command(result)
+            return result
+
+        return decorator
+
+    def group(self, cls=None, *args, **kwargs):
+        """A shortcut decorator that invokes :func:`.group` and adds it to
+        the internal command list via :meth:`~.GroupMixin.add_command`.
+        """
+        cls = None or Group
+
+        def decorator(func):
+            result = group(*args, **kwargs)(func)
+            self.add_command(result)
+            return result
+
+        return decorator
+
 
 class Group(Command, commands.Group):
     """Group command class for Red.
@@ -55,7 +135,28 @@ class Group(Command, commands.Group):
     This class inherits from `discord.ext.commands.Group`, with `Command` mixed
     in.
     """
-    pass
+
+    def __init__(self, *args, **kwargs):
+        self.autohelp = kwargs.pop("autohelp", False)
+        super().__init__(*args, **kwargs)
+
+    async def invoke(self, ctx):
+
+        view = ctx.view
+        previous = view.index
+        view.skip_ws()
+        trigger = view.get_word()
+        if trigger:
+            ctx.subcommand_passed = trigger
+            ctx.invoked_subcommand = self.all_commands.get(trigger, None)
+        view.index = previous
+        view.previous = previous
+
+        if ctx.invoked_subcommand is None or self == ctx.invoked_subcommand:
+            if self.autohelp and not self.invoke_without_command:
+                await ctx.send_help()
+
+        await super().invoke(ctx)
 
 
 # decorators
