@@ -36,9 +36,11 @@ import sys
 import traceback
 
 from . import commands
+from redbot.core.utils.chat_formatting import pagify, box
+from redbot.core.utils import fuzzy_command_search
 
 
-EMPTY_STRING = u"\u200b"
+EMPTY_STRING = "\u200b"
 
 _mentions_transforms = {"@everyone": "@\u200beveryone", "@here": "@\u200bhere"}
 
@@ -59,6 +61,12 @@ class Help(formatter.HelpFormatter):
         return isinstance(ctx.channel, discord.DMChannel)
 
     @property
+    def clean_prefix(self):
+        maybe_member = self.context.guild.me if self.context.guild else self.context.bot.user
+        pretty = f"@{maybe_member.display_name}"
+        return self.context.prefix.replace(maybe_member.mention, pretty)
+
+    @property
     def me(self):
         return self.context.me
 
@@ -70,12 +78,14 @@ class Help(formatter.HelpFormatter):
     def avatar(self):
         return self.context.bot.user.avatar_url_as(format="png")
 
-    @property
-    def color(self):
+    async def color(self):
         if self.pm_check(self.context):
-            return 0
+            return self.context.bot.color
         else:
-            return self.me.color
+            if await self.context.bot.db.guild(self.context.guild).use_bot_color():
+                return self.context.bot.color
+            else:
+                return self.me.color
 
     @property
     def destination(self):
@@ -121,11 +131,7 @@ class Help(formatter.HelpFormatter):
         """Formats command for output.
 
         Returns a dict used to build embed"""
-        emb = {
-            "embed": {"title": "", "description": ""},
-            "footer": {"text": self.get_ending_note()},
-            "fields": [],
-        }
+        emb = {"embed": {"title": "", "description": ""}, "footer": {"text": ""}, "fields": []}
 
         if self.is_cog():
             translator = getattr(self.command, "__translator__", lambda s: s)
@@ -143,6 +149,13 @@ class Help(formatter.HelpFormatter):
         if description:
             # <description> portion
             emb["embed"]["description"] = description[:2046]
+
+        tagline = await self.context.bot.db.help.tagline()
+        if tagline:
+            footer = tagline
+        else:
+            footer = self.get_ending_note()
+        emb["footer"]["text"] = footer
 
         if isinstance(self.command, discord.ext.commands.core.Command):
             # <signature portion>
@@ -177,22 +190,29 @@ class Help(formatter.HelpFormatter):
             for category, commands_ in itertools.groupby(data, key=category):
                 commands_ = sorted(commands_)
                 if len(commands_) > 0:
-                    field = EmbedField(category, self._add_subcommands(commands_), False)
-                    emb["fields"].append(field)
+                    for i, page in enumerate(
+                        pagify(self._add_subcommands(commands_), page_length=1000)
+                    ):
+                        title = category if i < 1 else f"{category} (continued)"
+                        field = EmbedField(title, page, False)
+                        emb["fields"].append(field)
 
         else:
             # Get list of commands for category
             filtered = sorted(filtered)
             if filtered:
-                field = EmbedField(
-                    "**__Commands:__**"
-                    if not self.is_bot() and self.is_cog()
-                    else "**__Subcommands:__**",
-                    self._add_subcommands(filtered),  # May need paginated
-                    False,
-                )
-
-                emb["fields"].append(field)
+                for i, page in enumerate(
+                    pagify(self._add_subcommands(filtered), page_length=1000)
+                ):
+                    title = (
+                        "**__Commands:__**"
+                        if not self.is_bot() and self.is_cog()
+                        else "**__Subcommands:__**"
+                    )
+                    if i > 0:
+                        title += " (continued)"
+                    field = EmbedField(title, page, False)
+                    emb["fields"].append(field)
 
         return emb
 
@@ -241,7 +261,7 @@ class Help(formatter.HelpFormatter):
         field_groups = self.group_fields(emb["fields"], page_char_limit)
 
         for i, group in enumerate(field_groups, 1):
-            embed = discord.Embed(color=self.color, **emb["embed"])
+            embed = discord.Embed(color=await self.color(), **emb["embed"])
 
             if len(field_groups) > 1:
                 description = "{} *- Page {} of {}*".format(
@@ -260,34 +280,31 @@ class Help(formatter.HelpFormatter):
 
         return ret
 
-    def simple_embed(self, ctx, title=None, description=None, color=None):
+    async def simple_embed(self, ctx, title=None, description=None, color=None):
         # Shortcut
         self.context = ctx
         if color is None:
-            color = self.color
+            color = await self.color()
         embed = discord.Embed(title=title, description=description, color=color)
         embed.set_footer(text=ctx.bot.formatter.get_ending_note())
         embed.set_author(**self.author)
         return embed
 
-    def cmd_not_found(self, ctx, cmd, color=None):
+    async def cmd_not_found(self, ctx, cmd, description=None, color=None):
         # Shortcut for a shortcut. Sue me
-        embed = self.simple_embed(
-            ctx,
-            title=ctx.bot.command_not_found.format(cmd),
-            description="Commands are case sensitive. Please check your spelling and try again",
-            color=color,
+        embed = await self.simple_embed(
+            ctx, title="Command {} not found.".format(cmd), description=description, color=color
         )
         return embed
 
-    def cmd_has_no_subcommands(self, ctx, cmd, color=None):
-        embed = self.simple_embed(
+    async def cmd_has_no_subcommands(self, ctx, cmd, color=None):
+        embed = await self.simple_embed(
             ctx, title=ctx.bot.command_has_no_subcommands.format(cmd), color=color
         )
         return embed
 
 
-@commands.command()
+@commands.command(hidden=True)
 async def help(ctx, *cmds: str):
     """Shows help documentation.
 
@@ -317,9 +334,19 @@ async def help(ctx, *cmds: str):
             command = ctx.bot.all_commands.get(name)
             if command is None:
                 if use_embeds:
-                    await destination.send(embed=ctx.bot.formatter.cmd_not_found(ctx, name))
+                    fuzzy_result = await fuzzy_command_search(ctx, name)
+                    if fuzzy_result is not None:
+                        await destination.send(
+                            embed=await ctx.bot.formatter.cmd_not_found(
+                                ctx, name, description=fuzzy_result
+                            )
+                        )
                 else:
-                    await destination.send(ctx.bot.command_not_found.format(name))
+                    fuzzy_result = await fuzzy_command_search(ctx, name)
+                    if fuzzy_result is not None:
+                        await destination.send(
+                            ctx.bot.command_not_found.format(name, fuzzy_result)
+                        )
                 return
         if use_embeds:
             embeds = await ctx.bot.formatter.format_help_for(ctx, command)
@@ -330,9 +357,17 @@ async def help(ctx, *cmds: str):
         command = ctx.bot.all_commands.get(name)
         if command is None:
             if use_embeds:
-                await destination.send(embed=ctx.bot.formatter.cmd_not_found(ctx, name))
+                fuzzy_result = await fuzzy_command_search(ctx, name)
+                if fuzzy_result is not None:
+                    await destination.send(
+                        embed=await ctx.bot.formatter.cmd_not_found(
+                            ctx, name, description=fuzzy_result
+                        )
+                    )
             else:
-                await destination.send(ctx.bot.command_not_found.format(name))
+                fuzzy_result = await fuzzy_command_search(ctx, name)
+                if fuzzy_result is not None:
+                    await destination.send(ctx.bot.command_not_found.format(name, fuzzy_result))
             return
 
         for key in cmds[1:]:
@@ -341,17 +376,27 @@ async def help(ctx, *cmds: str):
                 command = command.all_commands.get(key)
                 if command is None:
                     if use_embeds:
-                        await destination.send(embed=ctx.bot.formatter.cmd_not_found(ctx, key))
+                        fuzzy_result = await fuzzy_command_search(ctx, name)
+                        if fuzzy_result is not None:
+                            await destination.send(
+                                embed=await ctx.bot.formatter.cmd_not_found(
+                                    ctx, name, description=fuzzy_result
+                                )
+                            )
                     else:
-                        await destination.send(ctx.bot.command_not_found.format(key))
+                        fuzzy_result = await fuzzy_command_search(ctx, name)
+                        if fuzzy_result is not None:
+                            await destination.send(
+                                ctx.bot.command_not_found.format(name, fuzzy_result)
+                            )
                     return
             except AttributeError:
                 if use_embeds:
                     await destination.send(
-                        embed=ctx.bot.formatter.simple_embed(
+                        embed=await ctx.bot.formatter.simple_embed(
                             ctx,
                             title='Command "{0.name}" has no subcommands.'.format(command),
-                            color=ctx.bot.formatter.color,
+                            color=await ctx.bot.formatter.color(),
                         )
                     )
                 else:
