@@ -2,15 +2,15 @@ import pytest
 
 
 @pytest.fixture
-def mod(config):
+def mod(config, monkeypatch):
     from redbot.core import Config
 
-    Config.get_conf = lambda *args, **kwargs: config
+    with monkeypatch.context() as m:
+        m.setattr(Config, "get_conf", lambda *args, **kwargs: config)
+        from redbot.core import modlog
 
-    from redbot.core import modlog
-
-    modlog._register_defaults()
-    return modlog
+        modlog._register_defaults()
+        return modlog
 
 
 @pytest.mark.asyncio
